pyproject.toml
[project]
name = "agentic-ai-workflow-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipython>=9.12.0",
    "markdown>=3.10.2",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
